core/requirements.txt
pandas==1.2.4
numpy==1.20.1
fastapi==0.65.1
uvicorn==0.13.4
python-multipart==0.0.5
protobuf==3.20.1
scikit-learn==1.2.0
gunicorn==20.1.0
uvloop==0.16.0
httptools==0.3.0
psycopg2==2.8.6
python-dotenv==0.21.0
